# core/swagger/swagger.yaml
definitions:
  models.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  models.Rule:
    properties:
      location:
        type: string
      name:
        type: string
      status:
        type: boolean
      target_host:
        type: string
      target_port:
        type: string
      type:
        type: integer
    required:
    - name
    - status
    - target_host
    - target_port
    - type
    type: object
info:
  contact: {}
  description: Distributed Scan Services API DOCS
  license: {}
  title: Distributed Scan Services
  version: "1.0"
paths:
  /api/v1/grpc/client:
    get:
      consumes:
      - application/json
      parameters:
      - description: 模糊查询
        in: query
        name: search
        type: string
      - description: 当前页数,默认值:1
        in: query
        name: page
        type: string
      - description: 当前条数,默认值:10
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取gRPC客户端信息(IPV4、系统类型、系统版本、主机名、是否在线)
      tags:
      - Grpc
  /api/v1/rule:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 根据ID删除规则
      tags:
      - Rule
    get:
      consumes:
      - application/json
      parameters:
      - description: 模糊查询
        in: query
        name: search
        type: string
      - description: 枚举类型,1:单个IP类型 2:连续范围IP类型 3:网段类型
        in: query
        name: type
        type: integer
      - description: 状态查询,true/false
        in: query
        name: status
        type: string
      - description: 当前页数,默认值:1
        in: query
        name: page
        type: string
      - description: 当前条数,默认值:10
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取规则列表
      tags:
      - Rule
    post:
      consumes:
      - application/json
      parameters:
      - description: 规则请求体
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/models.Rule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 创建规则
      tags:
      - Rule
    put:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: query
        name: id
        required: true
        type: string
      - description: 规则请求体
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/models.Rule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 根据ID修改规则
      tags:
      - Rule
  /api/v1/rule/enum:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取规则所有枚举信息
      tags:
      - Rule
  /api/v1/scan:
    get:
      consumes:
      - application/json
      parameters:
      - description: 模糊查询
        in: query
        name: search
        type: string
      - description: 根据规则定义的location筛选扫描结果
        in: query
        name: location
        type: string
      - description: 根据日期(year-month-day)筛选扫描结果
        in: query
        name: date
        type: string
      - description: 当前页数,默认值:1
        in: query
        name: page
        type: string
      - description: 当前条数,默认值:10
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取扫描结果列表
      tags:
      - Scan
  /api/v1/scan/clear:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 作为计划任务每日执行,只保留最近7天扫描结果
      tags:
      - Scan
  /api/v1/scan/location:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取规则提供location作为查询参数
      tags:
      - Scan
  /api/v1/scan/remind:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 作为计划任务每日执行,对比今日与昨日是否有新增端口开放,若有则通过企业微信发送提醒
      tags:
      - Scan
  /api/v1/scan/trend:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 最近7天开放端口数量趋势
      tags:
      - Scan
  /api/v1/task:
    get:
      consumes:
      - application/json
      parameters:
      - description: 模糊查询
        in: query
        name: search
        type: string
      - description: 任务状态枚举,1:等待中,2:检测中,3:已完成,4:出错
        in: query
        name: status
        type: integer
      - description: 任务类型枚举,1:计划任务调度执行,2:手动执行
        in: query
        name: run_type
        type: integer
      - description: 当前页数,默认值:1
        in: query
        name: page
        type: string
      - description: 当前条数,默认值:10
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get 获取任务信息(执行时间、执行进度、执行状态、任务类型)
      tags:
      - Task
    post:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: query
        name: id
        required: true
        type: string
      - description: 任务类型枚举,1:计划任务调度执行,2:手动执行
        in: query
        name: run_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 执行扫描任务
      tags:
      - Task
  /api/v1/task/enum:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取任务所有枚举信息
      tags:
      - Task
swagger: "2.0"